% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bvnl}
\alias{bvnl}
\title{A C++ implementation of the bivariate normal CDF, derived from the MATLAB code of Alan Genz}
\usage{
bvnl(dh, dk, r)
}
\arguments{
\item{dh}{1st upper integration limit}

\item{dk}{2nd upper integration limit}

\item{r}{Correlation coefficient}
}
\description{
Calculates the probability that x < dh and y < dk
}
\details{
Original Matlab code written by: Alan Genz,
Department of Mathematics, Washington State University,
Pullman, Wa 99164-3113, Email: alangenz@wsu.edu,
}
\references{
Drezner, Z., & Wesolowsky, G. O. (1990). On the computation of the bivariate normal integral. Journal of Statistical Computation and Simulation, 35(1-2), 101-107.
}
\author{
David Gerard
}
