% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correst.R
\name{correst}
\alias{correst}
\title{Estimate the correlation between two SNPs from the genotype likelihoods and the
genotype distribution.}
\usage{
correst(X, Y, g, h, is_log = FALSE, method = c("mleCpp", "mleR"))
}
\arguments{
\item{X}{A matrix of genotype likelihoods. Element \code{(i, j)} is the
probability of the data for individual \code{i} given dosage \code{j - 1} for
SNP 1.}

\item{Y}{A matrix of genotype likelihoods. Element \code{(i, j)} is the
probability of the data for individual \code{i} given dosage \code{j - 1} for
SNP 2.}

\item{g}{A vector of probabilities. Element \code{j} is the probability
of dosage \code{j - 1} for SNP 1.}

\item{h}{A vector of probabilities. Element \code{j} is the probability
of dosage \code{j - 1} for SNP 2.}

\item{is_log}{A logical. Are \code{X}, \code{Y}, \code{g}, and \code{h} all on the
log-scale (\code{TRUE}) or not (\code{FALSE}).}

\item{method}{Which method should we use to estimate the correlation? MLE using
C++ (\code{"mleCpp"}) or  MLE using R (\code{"mleR"})?}
}
\value{
A list with the following elements:
\describe{
  \item{\code{zhat_gc}}{The estimate of the atanh of the correlation of the SNPs on the
      gaussian copula scale}
  \item{\code{sezhat_gc}}{The estimated standard error of zhat_gc.}
  \item{\code{corest}}{The estimated correlation of the SNPs on the original scale.}
  \item{\code{zhat}}{\code{atanh(corest)}}
  \item{\code{sezhat}}{The estimated standard error of zhat.}
}
}
\description{
Estimate the correlation between two SNPs from the genotype likelihoods and the
genotype distribution.
}
\author{
David Gerard
}
