% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correst.R
\name{correst_ind}
\alias{correst_ind}
\title{Estimate correlation using weights that assume independence.}
\usage{
correst_ind(pm1, pm2, g, h)
}
\arguments{
\item{pm1}{The posterior probability matrix for SNP 1. Element (i, j) is
the posterior probability that individual i has genotype j - 1.}

\item{pm2}{The posterior probability matrix for SNP 2. Element (i, j) is
the posterior probability that individual i has genotype j - 1.}

\item{g}{The pmf of SNP 1. ELement j is the posterior probability of genotype j - 1.}

\item{h}{The pmf of SNP 2. ELement j is the posterior probability of genotype j - 1.}
}
\description{
Estimate correlation using weights that assume independence.
}
\author{
David Gerard
}
