% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correst.R
\name{correst_ind_updog}
\alias{correst_ind_updog}
\title{Estimate correlation assuming independent weights using updog output.}
\usage{
correst_ind_updog(uout1, uout2)
}
\arguments{
\item{uout1}{A member of class \code{\link[updog]{flexdog}},
which is the output of fitting \code{flexdog} to one SNP.}

\item{uout2}{A member of class \code{\link[updog]{flexdog}},
which is the output of fitting \code{flexdog} to the other SNP.}
}
\description{
Estimate correlation assuming independent weights using updog output.
}
\author{
David Gerard
}
