% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correst.R
\name{correst_onestep_updog}
\alias{correst_onestep_updog}
\title{Estimate correlation using one Newton step from result of correst_ind_updog}
\usage{
correst_onestep_updog(uout1, uout2)
}
\arguments{
\item{uout1}{A member of class \code{\link[updog]{flexdog}},
which is the output of fitting \code{flexdog} to one SNP.}

\item{uout2}{A member of class \code{\link[updog]{flexdog}},
which is the output of fitting \code{flexdog} to the other SNP.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{zhat_gc}}{The estimate of the atanh of the correlation of
      the SNPs on the gaussian copula scale}
  \item{\code{sezhat_gc}}{The estimated standard error of zhat_gc.}
  \item{\code{corest}}{The estimated correlation of the SNPs on the
      original scale.}
  \item{\code{zhat}}{\code{atanh(corest)}}
  \item{\code{sezhat}}{The estimated standard error of zhat.}
}
}
\description{
Estimate correlation using one Newton step from result of correst_ind_updog
}
\author{
David Gerard
}
