% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corrlike}
\alias{corrlike}
\title{Likelihood of the correlation given the count data.}
\usage{
corrlike(atanh_rho, lX, lY, lg, lh)
}
\arguments{
\item{atanh_rho}{The inverse hyperbolic tangent of the correlation.
This is also known as the Fisher Z-transformation of the correlation.}

\item{lX}{A matrix with the genotype log-likelihoods for the first SNP. Element
(\code{i}, \code{j}) is the log-likelihood
of the data for individual \code{i} given dosage \code{j} for SNP 1.}

\item{lY}{A matrix with the genotype log-likelihoods for the second SNP. Element
(\code{i}, \code{j}) is the log-likelihood
of the data for individual \code{i} given dosage \code{j} for SNP 2.}

\item{lg}{A vector with the log-probabilities of the genotypes for the first SNP. Element
\code{j} is the log-probability of dosage \code{j - 1} in SNP 1.}

\item{lh}{A vector with the log-probabilities of the genotypes for the second SNP. Element
\code{j} is the log-probability of dosage \code{j - 1} in SNP 2.}
}
\description{
Likelihood of the correlation given the count data.
}
\seealso{
\describe{
  \item{\code{\link{correst}}}{For the function that optimizes \code{corrlike_r}.}
}
}
\author{
David Gerard
}
