% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_from_marg}
\alias{dist_from_marg}
\title{Get the probability mass function given marginal cdf's and the
correlation.}
\usage{
dist_from_marg(g_cdf, h_cdf, rho)
}
\arguments{
\item{g_cdf}{The cdf of the first SNP.}

\item{h_cdf}{The cdf of the second SNP.}

\item{rho}{The correlation.}
}
\value{
A matrix. The rows index the first SNP and the columns index the second SNP.
    Element (i, j) is the probability of genotype i - 1 in SNP 1 and j - 1 in SNP 2.
}
\description{
Get the probability mass function given marginal cdf's and the
correlation.
}
\author{
David Gerard
}
