% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_like}
\alias{gen_like}
\title{Calculate the genotypes log-likelihoods.}
\usage{
gen_like(refvec, sizevec, ploidy, seq, bias, od)
}
\arguments{
\item{refvec}{A vector of reference counts.}

\item{sizevec}{A vector of size counts.}

\item{ploidy}{The ploidy of the species.}

\item{seq}{The sequencing error rate.}

\item{bias}{The allele bias parameter.}

\item{od}{The overdispersion parameter.}
}
\value{
A matrix of numerics. The rows index the individuals and the
    columns index the genotype. Element (i, j) is the log-likelihood
    of dosage j for individual i.
}
\description{
Calculate the genotypes log-likelihoods.
}
\author{
David Gerard
}
