% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_cum_sum_exp}
\alias{log_cum_sum_exp}
\title{Log of cumulative sum of exponential.}
\usage{
log_cum_sum_exp(x)
}
\arguments{
\item{x}{A vector to log-cum-sum-exp.}
}
\value{
A NumericVector. Element i is the log of the sum of
    the exponential of the first i elements in \code{x}.
}
\description{
Log of cumulative sum of exponential.
}
\author{
David Gerard
}
