% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnormcop}
\alias{pnormcop}
\title{Distribution function of the bivariate Gaussian copula.}
\usage{
pnormcop(x, y, rho)
}
\arguments{
\item{x}{The value of the first element. Must be between 0 and 1.}

\item{y}{The value of the second element. Must be between 0 and 1.}

\item{rho}{The correlation. Must be between 0 and 1.}
}
\description{
Distribution function of the bivariate Gaussian copula.
}
\seealso{
\code{\link{bvnl}} For the cdf of the bivariate normal.
}
\author{
David Gerard
}
