% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uglycat.R
\docType{package}
\name{uglycat}
\alias{uglycat}
\alias{uglycat-package}
\title{\code{uglycat}: Using Genotype Likelihoods Yields Correlations and T-statistics}
\description{
\code{uglycat}: Using Genotype Likelihoods Yields Correlations and T-statistics
}
\section{\code{updog} Functions}{

\describe{
  \item{\code{\link{correst}}}{Estimate correlation.}
}
}

\author{
David Gerard
}
