% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correst.R
\name{correst}
\alias{correst}
\title{Estimate the correlation between two SNPs from the genotype likelihoods and the
genotype distribution.}
\usage{
correst(X, Y, g, h, is_log = FALSE)
}
\arguments{
\item{X}{A matrix of genotype likelihoods. Element \code{(i, j)} is the
probability of the data for individual \code{i} given dosage \code{j - 1} for
SNP 1.}

\item{Y}{A matrix of genotype likelihoods. Element \code{(i, j)} is the
probability of the data for individual \code{i} given dosage \code{j - 1} for
SNP 2.}

\item{g}{A vector of probabilities. Element \code{j} is the probability
of dosage \code{j - 1} for SNP 1.}

\item{h}{A vector of probabilities. Element \code{j} is the probability
of dosage \code{j - 1} for SNP 2.}

\item{is_log}{A logical. Are \code{X}, \code{Y}, \code{g}, and \code{h} all on the
log-scale (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Estimate the correlation between two SNPs from the genotype likelihoods and the
genotype distribution.
}
\author{
David Gerard
}
