% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correst.R
\name{correst}
\alias{correst}
\title{Estimate the correlation between two SNPs from the genotype likelihoods and the
genotype distribution.}
\usage{
correst(X, Y, g, h, is_log = FALSE, use_cpp = TRUE)
}
\arguments{
\item{X}{A matrix of genotype likelihoods. Element \code{(i, j)} is the
probability of the data for individual \code{i} given dosage \code{j - 1} for
SNP 1.}

\item{Y}{A matrix of genotype likelihoods. Element \code{(i, j)} is the
probability of the data for individual \code{i} given dosage \code{j - 1} for
SNP 2.}

\item{g}{A vector of probabilities. Element \code{j} is the probability
of dosage \code{j - 1} for SNP 1.}

\item{h}{A vector of probabilities. Element \code{j} is the probability
of dosage \code{j - 1} for SNP 2.}

\item{is_log}{A logical. Are \code{X}, \code{Y}, \code{g}, and \code{h} all on the
log-scale (\code{TRUE}) or not (\code{FALSE}).}

\item{use_cpp}{A logical Should we use the C++ implementation of the
optimization (\code{TRUE}) or not (\code{FALSE})? The fully C++ implementation
seems to be about 20 percent faster.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{zhat}}{The estimate of the atanh of the correlation of the SNPs on the
      transformed scale}
  \item{\code{se_zhat}}{The estimated standard error of zhat.}
  \item{\code{corest}}{The estimated correlation of the SNPs on the original scale.}
}
}
\description{
Estimate the correlation between two SNPs from the genotype likelihoods and the
genotype distribution.
}
\author{
David Gerard
}
